added chef installion conditional check in Jenkinsfile

+++ Jenkinsfile
@@ -8,10 +8,17 @@ pipeline {
     }
     stage ('Installing Chef Workstation') {
       steps {
-        sh 'sudo apt-get install -y wget tree unzip'
-        sh 'wget https://packages.chef.io/files/stable/chef-workstation/20.10.168/ubuntu/20.04/chef-workstation_20.10.168-1_amd64.deb'
-        sh 'sudo dpkg -i chef-workstation_20.10.168-1_amd64.deb'
-        sh 'sudo chef env --chef-license accept'
+        script {
+          def exists = fileExists '/usr/bin.chef-client'
+          if(exists == true) {
+            echo 'Skipping Chef Workstation install - already installed'
+          } else {
+            sh 'sudo apt-get install -y wget tree unzip'
+            sh 'wget https://packages.chef.io/files/stable/chef-workstation/20.10.168/ubuntu/20.04/chef-workstation_20.10.168-1_amd64.deb'
+            sh 'sudo dpkg -i chef-workstation_20.10.168-1_amd64.deb'
+            sh 'sudo chef env --chef-license accept'
+          }
+        }
       }
     }
     stage ('Third stage') {
